sidecar/requirements.txt
# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0

# HTTP
httpx>=0.28.0
aiofiles>=24.1.0

# Image processing
Pillow>=11.0.0
opencv-python-headless>=4.10.0

# AI — Vision (choose one or both)
openai>=1.50.0           # GPT-4o Vision (API key required)
# torch>=2.5.0           # Uncomment for local PyTorch models
# torchvision>=0.20.0
# transformers>=4.46.0   # CLIP / ViT local inference
# ultralytics>=8.3.0     # YOLOv8 piece detection (heavy, ~500MB)

# Utilities
numpy>=1.26.0
